Proper namespacing of private library classes/modules, proper gem folder structure, tidy up requires.
