+++ lib/contracts/builtin_contracts.rb
@@ -1,4 +1,4 @@
-require 'testable'
+require 'contracts/testable'
 
 =begin rdoc
 This module contains all the builtin contracts.
