+++ lib/contracts/testable.rb
